# docs/openAPI_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Expense SaaS API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/signup": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Signup",
        "description": "Register a new user, create a default tenant and return auth tokens.\n\nArgs:\n    payload: SignupIn input schema with email, password and optional name.\n    db: Async DB session.\n\nReturns:\n    TokenOut containing an access token and (in test mode) a refresh token.\n\nRaises:\n    HTTPException 400 when email already exists.",
        "operationId": "signup_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "Authenticate a user and issue access/refresh tokens.\n\nArgs:\n    payload: LoginIn schema with email and password.\n    db: Async DB session.\n\nReturns:\n    TokenOut containing access token and optional refresh token (in test mode).\n\nRaises:\n    HTTPException 401 when credentials are invalid.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh",
        "description": "Rotate a refresh token and issue a new access token.\n\nArgs:\n    refresh_token: Opaque refresh token string presented by the client.\n    db: Async DB session.\n\nReturns:\n    TokenOut containing new access token and (in test mode) the new refresh token.\n\nRaises:\n    HTTPException 401 when the refresh token is invalid or expired.",
        "operationId": "refresh_auth_refresh_post",
        "parameters": [
          {
            "name": "refresh_token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "description": "Revoke a refresh token to log a user out.\n\nArgs:\n    refresh_token: Opaque refresh token to revoke.\n    db: Async DB session.\n\nReturns:\n    Dict indicating operation success.",
        "operationId": "logout_auth_logout_post",
        "parameters": [
          {
            "name": "refresh_token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/tenants/{tenant_id}/invite": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create Invite",
        "description": "Create an invite record for a tenant (placeholder implementation).\n\nNotes:\n    This is a simplified flow that does not enforce membership/role checks.\n    In production use, wire get_current_user and require_membership to verify permissions.\n\nArgs:\n    tenant_id: Tenant identifier to invite into.\n    payload: InviteCreate schema with invitee email and role.\n    db: Async DB session.\n\nReturns:\n    A small hint of the invite token that was generated.",
        "operationId": "create_invite_auth_tenants__tenant_id__invite_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "args",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Args"
            }
          },
          {
            "name": "kwargs",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Kwargs"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InviteCreate": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MembershipRole"
              },
              {
                "type": "null"
              }
            ],
            "default": "member"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "InviteCreate",
        "description": "Input schema to create an invite for a tenant.\n\nArgs:\n    email: Email address to invite.\n    role: Role to assign when the invite is accepted."
      },
      "LoginIn": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "tenant_uuid": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Uuid"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginIn",
        "description": "Input schema for user login.\n\nArgs:\n    email: User email address.\n    password: Plaintext password.\n    tenant_uuid (Option): specify a tenant to login"
      },
...
      "MembershipRole": {
        "type": "string",
        "enum": [
          "owner",
          "member",
          "viewer"
        ],
        "title": "MembershipRole",
        "description": "Role assigned to a membership within a tenant.\n\nOWNER has full control, MEMBER has standard access, VIEWER has read-only access."
      },
...
      "SignupIn": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "SignupIn",
        "description": "Input schema for user signup.\n\nArgs:\n    email: User email address used for login.\n    password: Plaintext password provided by the user (will be hashed).\n    name: Optional display name."
      },
...
      "TokenOut": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "TokenOut",
        "description": "Output schema for authentication tokens.\n\nAttributes:\n    access_token: JWT access token string.\n    token_type: Type of token (default: 'bearer').\n    refresh_token: Optional opaque refresh token (only present in tests)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}